Initial commit.
Added basic title and started working on basic room capabilities. Good next step would be to expand this and add the ability to change the room based on it's id. The whole idea here is to be able to read the room definitions from a file. As we go along, we can add additional information to the Room struct, such as items, enemies, and special events.