Basic doors kind of working.
Might want to switch your approach here and have doors toggle
open/closed. You can create and remove an object for the door as needed.
Would reduce a lot of the spread out logic for doors.

Downside: rendering order needs to be considered, but that might not be
too difficult, and you should probably have some control over that
anyways.